requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
appdirs==1.4.4
    # via fs
click==8.1.7
fs==2.4.16
numpy==1.26.4
    # via rawpy
pillow==10.3.0
rawpy==0.21.0
setuptools==69.5.1
    # via fs
six==1.16.0
    # via fs
